Add files via upload
This network architecture diagram consists out of 3 subnets. 

Red Private Subnet:
On the left in the red square, there is a private subnet that supports up to 30 hosts. the local router in this subnet is directly connected to Router in the blue private subnet. Router in the red square is also directly connected to the router in the public subnet, the router in the LAN has no direct connectivity to the public internet. It has to go through the internet gateway and NAT gateway before it reaches the Internet.

Blue Private Subnet with NAT Gateway:
The blue subnet supports 62 hosts and contains a router and NAT gateway. The Internet gateway is directly connected to the NAT gateway in the blue public subnet, the NAT gateway has a direct link to the internet through the Internet Gateway.

The Green subnet supports up to 5 hosts and contains a router and an Internet Gateway. This router is directly connected to the router at the blue subnet and provides Internet connectivity.. Red and green subnets are also connected with each other through the routers. Each subnet has a switch whereby the hosts can communicate with each other. The blue bricks are illustrating the firewalls between the Internet and the subnets and between the networks. The Network Address Translation gives the network access to the Internet without having a public IP address.